This error has popped up every time
I actually went to Professor Vong and Frank with this error, and although there is a temporary solution to solving it (restarting my computer), the error persists anyways. This happens to my Windows laptop. I work mainly on my Mac anyways, but my Mac takes a really long time to update anything I change in the code on the website itself. For example, it'll show the same error that I fixed up in the frontend code without any changes that I actually made to it.