Collab Feature LLD
I actually did make an LLD for when I was redesigning the Collab Feature after Shane dropped (in the middle of working on Collab Search), but I just didn't upload it. This was made weeks ago. The markdown of the mermaid and the png file are attached above. I was in the middle of testing out the Success Request case, but ran into errors and got caught up fixing those. That's probably why I never uploaded this.